Colocate contributor auth logic in isAuthorizedContributor action
Remove ContributorAuthorization lib; export helpers from the action entry
for unit tests. Drop CONTRIBUTOR_PLUS_TEAM_SLUG from CONST.
